Move stateful EOS integration tests to a separate file
